requirements.txt
beem
google-generativeai
requests
